Don't warn when kotlin projects redundantly set the allowed usage on the test fixtures runtime configuration
The user can't do anything about these warnings, so suppress them.  This was a case missed during the original work on this feature.

This commit also refactors the `AbstractJavaTestFixturesIntegrationTest` class to create a deeper hierarchy, containing a base class for Java project-based test fixture tests and a new top-level class for Kotlin project-based test fixture tests.
